// driversupport_v2-rb_0.1.95/src/app/shared/models/user.model.ts
export interface CurrentUser {
  app: string;
  depots: Depot[];
  roles: string[];
  user: string;
  total: number;
}

export interface User {
  app: string;
  depot: Depot;
  roles: string[];
  user: string;
  total: number;
}

export interface Depot {
  bezeichnung: string;
  code: string;
  depotnr: string;
  id: number;
}